backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.12.0
opencv-python==4.9.0.80
torch==2.2.1
pandas==2.2.3
flask-cors==5.0.1
pymongo==4.11.2
scikit-learn==1.6.1
opensmile==2.5.1
PyJWT~=2.10.1
werkzeug~=3.1.3
flask-session~=0.8.0
PyInstaller==6.3.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0